Moving colors
The goal is to avoid too much repetition. The prompt (on the left) is
already yellow and the nvim `normal` mode (on the right) is already blue
-- so swapping the date and window colors helps a lot. But nvim's blue
tabs are also on the left, so making the window purple is nicer. We
don't have a diff_purple color and making one is annoying -- but
diff_blue works well enough, much like diff_yellow works with orange.