Tweaking Harpoon usage
Going back to adding files on :w. To avoid cluttering, only up to 5
items can be in the list at any time. Using :A pins the file: moves it
at the top of the list and doesn't count towards the :w limit. So we
have the 5 most recent files on top of whatever was explicitly marked.