Appending perm hooks
We're prepending temp hooks so that the most active buffers are to the
left which is nicer than being in the corner. But prepending perm hooks
means changing the order of the other perms which is bad.

So we append perm, and prepend temp. (Actually we prepend everything,
but then we flip the perms when using them for navigation.)