Don't re-hook temp files
Speaking of keeping the order: rehooking temp files changes the order of
everything and it made it borderline unusable in practice. So I'm just
naively removing that mechanic. In practice this means losing the "most
recent files" behavior, because when going over 9, the next file to be
removed is not necessarily the oldest one.

So conceptually it's not great, but let's see if in practice if feels
better than the alternative.